Features/filters

+++ backend/Application/UseCases/CreateCrime/CreateCrimeRequest.cs
@@ -8,6 +8,6 @@
         DateTime WantedPersonBirthDate,
         DateTime CrimeDate,
         string Location,
-        decimal PointLatitude,
-        decimal PointLongitude);
+        double PointLatitude,
+        double PointLongitude);
 }
