+++ backend/Application/UseCases/GetAllCrimes/ShowOnMapCrimeResponse.cs
@@ -1,5 +1,12 @@
 ﻿namespace Application.UseCases.GetAllCrimes
 {
-    public record ShowOnMapCrimeResponse(Guid Id, Guid CrimeTypeId, string? Location, DateTime CrimeDate, decimal PointLatitude, decimal PointLongitude);
+    public record ShowOnMapCrimeResponse(
+        Guid Id,
+        Guid CrimeTypeId,
+        string? Location,
+        DateTime CrimeDate,
+        double PointLatitude,
+        double PointLongitude
+    );
 
 }
